% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/adobecolor-package.R
\docType{package}
\name{adobecolor}
\alias{adobecolor}
\alias{adobecolor-package}
\title{A package to read (and eventually write) Adobe color/swatch files (.aco / .ase)}
\description{
A package to read (and eventually write) Adobe color/swatch files (.aco / .ase)
}
\author{
Bob Rudis (@hrbrmstr)
}

