% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/adobecolor.R
\name{read_aco}
\alias{read_aco}
\title{Read colors from Adobe Color (ACO) files}
\usage{
read_aco(path, use_names = TRUE, .verbose = TRUE)
}
\arguments{
\item{path}{partial or full file path or URL to an ACO file}

\item{use_names}{add color names to the vector (defaults to \code{TRUE}). See NOTE}

\item{.verbose}{show extra information about ACO file processing#'}
}
\description{
Given a path or URL to an \code{.aco} file, this function will return
a named character vector (if color names are present) of hex RGB colors.
}
\note{
When using named color palettes in a \code{ggplot2} \code{scale_} context, you
    must \code{unname} them or set \code{use_names} to \code{FALSE}.
    Not sure if this is a bug or a deliberate feature in ggplot2. Also, Neither Lab nor
    greyscale colors are supported.
}
\examples{
\dontrun{
# built in palette
eighties <- read_ase(system.file("palettes", "tomorrow_night_eighties.aco", package="adobecolor"))
print(eighties)
show_palette(eighties)

# from the internet directly
url <- "https://github.com/chriskempson/tomorrow-theme/raw/master/Photoshop\%20Color\%20Swatchesdi/Tomorrow\%20Night.aco"
tomorrow_night <- read_ase(url)
print(tomorrow_night)
show_palette(tomorrow_night)
}
}

